update pr comments
