+++ packages/contact-center/cc-components/src/components/task/CampaignTask/CampaignTaskPopover/campaign-task-popover.tsx
@@ -2,8 +2,7 @@ import React, {useRef} from 'react';
 import {Popover} from '@momentum-design/components/dist/react';
 import CampaignTaskListItem from '../CampaignTaskListItem/campaign-task-list-item';
 import GlobalVariablesPanel from '../../GlobalVariablesPanel/global-variables-panel';
-import {CampaignTaskPopoverProps} from './campaign-task-popover.types';
-import {CallAssociatedDataMap, getCallerIdentifier} from '../../task.types';
+import {CampaignTaskPopoverProps, CallAssociatedDataMap, getCallerIdentifier} from '../../task.types';
 import {getAgentViewableGlobalVariables} from '../../Task/task.utils';
 import {getCampaignCpd} from '../../TaskList/task-list.utils';
 import './campaign-task-popover.style.scss';
