+++ packages/contact-center/cc-components/src/components/task/task.types.ts
@@ -13,6 +13,7 @@ import {
   AddressBookEntrySearchParams,
   AddressBookEntriesResponse,
 } from '@webex/cc-store';
+import {CampaignErrorType} from './CampaignErrorDialog/campaign-error-dialog.types';
 
 type Enum<T extends Record<string, unknown>> = T[keyof T];
 
@@ -921,4 +922,258 @@ export interface ButtonConfig {
   className: string;
   disabled?: boolean;
   isVisible: boolean;
+}
+
+// ==================== GLOBAL VARIABLES PANEL TYPES ====================
+
+/**
+ * Properties for the GlobalVariablesPanel component.
+ */
+export interface GlobalVariablesPanelProps {
+  /**
+   * List of agent-viewable global variables to display.
+   */
+  variables: CADVariable[];
+
+  /**
+   * Optional CSS class name for additional styling.
+   */
+  className?: string;
+
+  /**
+   * Layout mode for the variables grid.
+   * - `single-column`: one variable per row (used in the inline card)
+   * - `two-column`: two variables per row (used in the popover)
+   * @default 'single-column'
+   */
+  layout?: 'single-column' | 'two-column';
+
+  /**
+   * CSS background value for the panel container.
+   * @default 'var(--mds-color-theme-background-glass-normal)'
+   */
+  panelBackground?: string;
+}
+
+// ==================== CAMPAIGN TYPES ====================
+
+/**
+ * Auto-action to perform when the campaign preview offer times out.
+ * Matches the values from callProcessingDetails.campaignPreviewAutoAction.
+ */
+export type CampaignAutoAction = 'ACCEPT' | 'SKIP' | 'REMOVE';
+
+/**
+ * Maps a CampaignAutoAction to the corresponding CampaignErrorType
+ * used when the auto-action or manual action fails.
+ */
+export const CAMPAIGN_ACTION_ERROR_MAP: Record<CampaignAutoAction, CampaignErrorType> = {
+  ACCEPT: 'ACCEPT_FAILED',
+  SKIP: 'SKIP_FAILED',
+  REMOVE: 'REMOVE_FAILED',
+};
+
+/**
+ * Keys of CAMPAIGN_ACTION_ERROR_MAP — used to type the error handler in CampaignTask.
+ */
+export type CampaignErrorActionType = keyof typeof CAMPAIGN_ACTION_ERROR_MAP;
+
+/**
+ * Campaign-specific fields on `callProcessingDetails`.
+ *
+ * These fields are present at runtime on campaign preview reservation
+ * events but are not yet part of the installed SDK type definitions.
+ * This bridge type can be removed once the SDK package is updated.
+ */
+export interface CampaignCallProcessingDetails {
+  /** Campaign name (not UUID) */
+  campaignId?: string;
+  /** Campaign type (e.g. 'preview_standard', 'preview_direct') */
+  campaignType?: string;
+  /** Indicates if the skip action is disabled for campaign preview contacts */
+  campaignPreviewSkipDisabled?: string;
+  /** Indicates if the remove action is disabled for campaign preview contacts */
+  campaignPreviewRemoveDisabled?: string;
+  /** Auto-action to perform when campaign preview offer times out (ACCEPT, SKIP, REMOVE) */
+  campaignPreviewAutoAction?: string;
+  /** Timestamp (ms) when the campaign preview offer expires */
+  campaignPreviewOfferTimeout?: string;
+}
+
+/**
+ * Properties for the CampaignTask component.
+ *
+ * The component renders campaign preview contact details, action buttons
+ * (Accept / Skip / Remove), a countdown timer, and an error dialog.
+ * When the countdown expires the configured auto-action is triggered.
+ *
+ * Following the pattern used by the Task component, SDK operations are
+ * passed in as callback props rather than passing the cc instance directly.
+ */
+export interface CampaignTaskProps {
+  /**
+   * The campaign preview task (AgentOfferCampaignReservation).
+   * Campaign metadata is read from `task.data.interaction.callProcessingDetails`.
+   */
+  task: ITask;
+
+  /**
+   * Accepts the campaign preview contact and initiates the outbound call.
+   * Wraps `cc.acceptPreviewContact({ interactionId, campaignId })`.
+   */
+  acceptPreviewContact: () => Promise<void>;
+
+  /**
+   * Skips the campaign preview contact and moves to the next one.
+   * Wraps `cc.skipPreviewContact({ interactionId, campaignId })`.
+   */
+  skipPreviewContact: () => Promise<void>;
+
+  /**
+   * Removes the campaign preview contact from the campaign list.
+   * Wraps `cc.removePreviewContact({ interactionId, campaignId })`.
+   */
+  removePreviewContact: () => Promise<void>;
+
+  /**
+   * Cancels the campaign preview call by ending the task.
+   * Wraps `task.end()`.
+   */
+  cancelPreviewContact: () => Promise<void>;
+
+  /**
+   * Whether the agent is logged in with a Browser (WebRTC) device.
+   * When true the Cancel button is rendered so the agent can end the
+   * WebRTC call.  For AGENT_DN the phone handles hangup, so the Cancel
+   * button is hidden — consistent with Agent Desktop behaviour.
+   */
+  isBrowser?: boolean;
+
+  /**
+   * Logger instance for logging purposes.
+   */
+  logger?: ILogger;
+
+  /**
+   * Whether this campaign preview has been accepted.
+   * Driven by the store's `acceptedCampaignIds` set — survives component
+   * remounts caused by transient task-list updates during the accept
+   * transition.  When `true`, action buttons and countdown are hidden
+   * and the handle-time timer is shown instead.
+   */
+  isAccepted?: boolean;
+
+  /**
+   * The logged-in agent's ID.  Used to look up the agent's participant
+   * entry when reading `joinTimestamp` for the handle-time timer.
+   */
+  agentId?: string;
+}
+
+/**
+ * Properties for the CampaignTaskPopover component.
+ *
+ * Displays a hover popover over the campaign preview task with the
+ * ListItem row (avatar, title, phone, countdown, action buttons)
+ * and a two-column scrollable data panel of global variables.
+ */
+export interface CampaignTaskPopoverProps {
+  /** The campaign preview task. */
+  task: ITask;
+
+  /** Logger instance for logging purposes. */
+  logger?: ILogger;
+
+  /** ID of the trigger element that opens the popover on hover. */
+  triggerId: string;
+
+  /** Whether the Accept button has been clicked (shows "Connecting..." state). */
+  isAcceptClicked: boolean;
+
+  /** Whether the campaign preview has been accepted by the backend (call controls visible). */
+  isAccepted: boolean;
+
+  /** Whether the Accept button is disabled. */
+  isAcceptDisabled: boolean;
+
+  /** Whether the Skip button is disabled. */
+  isSkipDisabled: boolean;
+
+  /** Whether the Remove button is disabled. */
+  isRemoveDisabled: boolean;
+
+  /** Handler for Accept button click. */
+  onAccept: () => void;
+
+  /** Handler for Skip button click. */
+  onSkip: () => void;
+
+  /** Handler for Remove button click. */
+  onRemove: () => void;
+
+  /** Handler for countdown timeout. */
+  onTimeout: () => void;
+
+  /** Timestamp (ms) when the campaign call was accepted — used for the handle time timer. */
+  handleTimestamp?: number;
+}
+
+/**
+ * Properties for the CampaignTaskListItem component.
+ *
+ * Renders the ListItem row shared between the CampaignTask card
+ * and CampaignTaskPopover: avatar, title, phone, countdown, and
+ * Accept / Skip / Remove action buttons.
+ */
+export interface CampaignTaskListItemProps {
+  /** Display title (customer name or caller identifier). */
+  title: string;
+
+  /** Phone number to show as secondary label. */
+  phoneNumber?: string;
+
+  /** Customer name — used to decide whether to show phone as secondary label. */
+  customerName?: string;
+
+  /** Campaign preview offer timeout timestamp (ms string). */
+  timeoutTimestamp?: string;
+
+  /** Whether the Accept button has been clicked (shows "Connecting..." state). */
+  isAcceptClicked: boolean;
+
+  /** Whether the campaign preview has been accepted by the backend (call controls visible). */
+  isAccepted: boolean;
+
+  /** Whether the Accept button is disabled. */
+  isAcceptDisabled: boolean;
+
+  /** Whether the Skip button is disabled. */
+  isSkipDisabled: boolean;
+
+  /** Whether the Remove button is disabled. */
+  isRemoveDisabled: boolean;
+
+  /** Handler for Accept button click. */
+  onAccept: () => void;
+
+  /** Handler for Skip button click. */
+  onSkip: () => void;
+
+  /** Handler for Remove button click. */
+  onRemove: () => void;
+
+  /** Handler for countdown timeout. */
+  onTimeout: () => void;
+
+  /** Timestamp (ms) when the campaign call was accepted — used for the handle time timer. */
+  handleTimestamp?: number;
+
+  /** Logger instance. */
+  logger?: ILogger;
+
+  /** Optional CSS class name applied to the ListItem. */
+  className?: string;
+
+  /** Optional test ID prefix for data-testid attributes. */
+  testIdPrefix?: string;
 }
